Recommend gnome-session-xsession rather than hard-require it
This makes it possible to avoid installing gnome-session-xsession, which
in turn pulls in xorg-x11-server-Xorg. It's still going to be pulled in
to the default install as Recommends are installed by default.

The motivation for changing this is that there are some arm devices that
only work with Wayland and where Xorg is unsupported, as requested by
pbrobinson.